requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pyyaml>=6.0
python-dotenv>=1.0
pytz>=2024.1
tenacity>=8.2
python-dateutil>=2.9
markdown>=3.5
apscheduler>=3.10,<4.0
tzdata>=2024.1
